fix string on linux (#4)
* cleanup

* start replace type to wstring

* add string_proxy

* tests nanodbc_c result statement connection and database_metadata

* fix wstring

+++ cpp/src/core/isolation_level.cpp
@@ -25,7 +25,7 @@ int32_t IsolationLevel::to_odbc() const {
 
 /// \brief Convert to integer for C API compatibility.
 IsolationLevel::operator int() const {
-    return static_cast<int>(to_odbc());
+    return to_odbc();
 }
 
 // Map ODBC constant back to our enum
